refactor to fix enum-related bugs
